+++ docs/changelog.md
@@ -20,6 +20,7 @@ pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Am
 - Add datetime to each logging operation.
 - Adapt defacl query in order to enhance compatibility.
 - Ensure password from YAML are string.
+- Don't remove member before dropping them.
 - API: Default to current os.environ.
 
 Support for running ldap2pg on Python 2.7 is now deprecated.